Allow for partitioning of basic and counting BFs.
This patch adds a new parameter to the constructor of basic and counting Bloom
filters to enable *partitioning*, that is, using each of the *k* hash functions
locally within a partition of the underlying vector of bits/cells.

This allows, for example, creating a Count-Min Sketch simply by enabling
partitioning on a counting Bloom filter.

The bf test utility had an unused parameter 'p' to enable partitioning. It is
now enabled for basic and counting Bloom filters.